+++ tests/TestCase/BodyParserTest.php
@@ -28,8 +28,33 @@ class BodyParserTest extends TestCase
     use ReflectionTrait;
 
     /**
-     * Test for `extractLinks()` method
+     * @var int
+     */
+    protected static int $currentErrorLevel;
+
+    /**
+     * @inheritDoc
+     */
+    public static function setUpBeforeClass(): void
+    {
+        parent::setUpBeforeClass();
+
+        self::$currentErrorLevel = error_reporting(E_ALL & ~E_USER_DEPRECATED);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public static function tearDownAfterClass(): void
+    {
+        parent::tearDownAfterClass();
+
+        error_reporting(self::$currentErrorLevel);
+    }
+
+    /**
      * @test
+     * @uses \Tools\BodyParser::extractLinks()
      */
     public function testExtractLinks(): void
     {