+++ tests/TestCase/GlobalFunctionsTest.php
@@ -18,7 +18,6 @@ namespace Tools\Test;
 use App\ExampleChildClass;
 use App\ExampleClass;
 use App\ExampleOfStringable;
-use LogicException;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 use Tools\TestSuite\TestTrait;
@@ -42,9 +41,8 @@ class GlobalFunctionsTest extends TestCase
         //This class implements the `__toString()` method
         $this->assertSame('[\'a\', \'App\ExampleOfStringable\']', array_to_string(['a', new ExampleOfStringable()]));
 
-        foreach ([['a', true], ['a', ['b', 'c']]] as $array) {
-            $this->assertException(fn() => array_to_string($array), LogicException::class, 'Cannot convert array to string, some values are not stringable');
-        }
+        $this->expectExceptionMessage('Cannot convert array to string, some values are not stringable');
+        array_to_string(['a', ['b', 'c']]);
     }
 
     /**