+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 1.x branch
 ## 1.8 branch
+### 1.8.2
+* added tests for PHP 8.2;
+* `BodyParser` is deprecated and will be removed in a later release;
+* `TestTrait::assertException()` is deprecated and will be removed in a later release;
+* `TestTrait::expectAssertionFailed()` is deprecated and will be removed in a later release.
+
 ### 1.8.1
 * tests for `ReflectionTrait` and `TestTrait` have largely improved;
 * the `TestCase` class has been deprecated and will be removed in a later release. Use instead the